+++ .env
@@ -1,0 +1,2 @@
+VITE_BASE_URL =https://restcountries.com/v3.1
+VITE_COHERE_TOKEN= Mvm7OajHkGHzQ3s22eLgn3sPeEHUdRN23Wss0mfN
