+++ src/App.tsx
@@ -1,18 +1,18 @@
 import { createBrowserRouter } from "react-router-dom"
-import MainLayout from "./layouts/Main.layout"
+// import MainLayout from "./layouts/Main.layout"
 import HomeLayout from "./layouts/Home.Layout"
 import { AppRoutes } from "./types/routes"
 
 import ErrorPage from "./modules/ErrorPage"
 
-import GetStartedPage from "./pages/GetStarted"
-import HomePage from "./pages/Home"
+import GetStartedPage from "./pages/Home"
+import HomePage from "./pages/Countries"
 import DetailPage from "./pages/Detail"
 
 export const router = createBrowserRouter([
   {
     path: AppRoutes.getStarted,
-    element: <MainLayout />,
+    element: <HomeLayout />,
     errorElement:<ErrorPage/>,
     children: [
       {
