+++ src/components/Header/index.tsx
@@ -1,32 +1,17 @@
+import { useCallback, useEffect, useState } from "react";
 import { Link } from "react-router-dom";
-import { AppRoutes } from "../../types/routes";
 import { MdOutlineDarkMode } from "react-icons/md";
 import { FaGithub } from "react-icons/fa";
-import { useCallback, useEffect, useState } from "react";
 import { FiSun } from "react-icons/fi";
-import { FaDesktop } from "react-icons/fa6";
+import { AppRoutes } from "../../types/routes";
 
-export default function Index(): JSX.Element {
+export default function Index() {
   const [theme, setTheme] = useState<string | null>(
     localStorage.getItem("theme") ? localStorage.getItem("theme") : "system"
   );
   const element = document.documentElement;
   const darkQuery = window.matchMedia("(prefers-color-scheme:dark)");
 
-  const iconSet = [
-    {
-      icon: <FiSun />,
-      title: "light",
-    },
-    {
-      icon: <MdOutlineDarkMode />,
-      title: "dark",
-    },
-    {
-      icon: <FaDesktop />,
-      title: "system",
-    },
-  ];
 
   const onWindowMatch = useCallback(() => {
     if (
@@ -68,11 +53,23 @@ export default function Index(): JSX.Element {
     }
   });
 
+  const toggleTheme = () => {
+    if (theme === "dark") {
+      setTheme("light");
+    } else {
+      setTheme("dark");
+    }
+  };
+
+  const isDark =
+    theme === "dark" ||
+    (theme === "system" && darkQuery.matches);
+
   return (
-    <header className="fixed bg-white top-0 left-0 h-16 shadow-sm  w-full z-20 dark:bg-bgDark">
+    <header className="fixed bg-white top-0 left-0  w-full z-20 dark:bg-bgDark">
       <div className="px-4 py-2 sm:px-12 pb-4">
         <nav>
-          <div className="flex items-center justify-between mx-auto py-1">
+          <div className="flex items-center justify-between mx-auto  px-8  py-4 max-w-screen-xl rounded-full shadow-md  bg-white dark:bg-bgDark">
             <div>
               <Link
                 to={AppRoutes.home}
@@ -87,40 +84,13 @@ export default function Index(): JSX.Element {
                   <FaGithub />
                 </a>
               </div>
-              <div className="bg-textWhite  rounded-md px-2">
-                {iconSet.map((icon) => (
-                  <button
-                    className={`w-8 h-8 leading-9 text-2xl rounded-xl ${
-                      theme === icon.title && "text-primary"
-                    } m-1`}
-                    onClick={() => setTheme(icon.title)}
-                  >
-                    <span key={icon.title}>{icon.icon}</span>
-                  </button>
-                ))}
-              </div>
+              <button
+                onClick={toggleTheme}
+                className="w-8 h-8 text-2xl rounded-xl text-black dark:text-primary"
+              >
+                {isDark ? <FiSun /> : <MdOutlineDarkMode />}
+              </button>
             </div>
-          </div>
-          <div className="image-container">
-            <picture className="flex justify-end items-center">
-              <source srcSet="/src/assets/bgAvif.avif" type="image/avif" />
-              <img
-                src="/src/assets/lightBg.png"
-                alt=""
-                className="w-[71.75rem] flex-none max-w-none dark:hidden"
-                decoding="async"
-              />
-            </picture>
-
-            <picture className="flex justify-end items-center">
-              <source srcSet="/src/assets/dark.avif" type="image/avif" />
-              <img
-                src="/src/assets/darkBg.png"
-                alt=""
-                className="w-[90rem] flex-none max-w-none z-50 hidden dark:block"
-                decoding="async"
-              />
-            </picture>
           </div>
         </nav>
       </div>
